Move a much infrastructure as possible to helper mod
